car_customization.py
# car_customization.py
import openai
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Function to get customization suggestions based on transcription
def get_customization_suggestions(transcription):
    # Set up the OpenAI API key (if needed, update this with the correct logic)
    openai.api_key = os.getenv("OPENAI_API_KEY")
    
    # Call OpenAI API to generate customization suggestions (modify as needed)
    response = openai.Completion.create(
        engine="text-davinci-003",  # Or any model of your choice
        prompt=f"Given the transcription: '{transcription}', provide car customization suggestions.",
        max_tokens=150
    )
    
    return response.choices[0].text.strip()

# Function to generate car image (optional)
def generate_car_image(description):
    # Use DALL-E or other image generation logic here (if needed)
    pass
