requirements.txt
whisper
streamlit
python-dotenv
openai
